requirements.txt
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.2.0; python_full_version >= '3.7.0'
idna==3.4; python_version >= '3.5'
imgurpython==1.1.7
python-dotenv==1.0.0
requests==2.31.0; python_version >= '3.7'
urllib3==2.0.4; python_version >= '3.7'